app/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
langchain-openai
langchain-core
langchain-milvus
langchain-text-splitters
langchain-unstructured
pydantic==2.6.4
unstructured[pdf]
python-magic==0.4.27
python-dotenv==1.0.1
